Fix overflow: use ll instead of int for area calculation

+++ content/geometry/lattice-point.cpp
@@ -3,15 +3,15 @@
  - Only works for integers points
  *Status:* Tested on CSES
 */
-pair<int, int> latticePoints(vector<Point2D<int>> &P) {
+pair<ll, ll> latticePoints(vector<Point2D<ll>> &P) {
   P.push_back(P.front());
-  int area = 0, bounds = 0;
-  for(int i = 0; i < P.size()-1; ++i) {
+  ll area = 0, bounds = 0;
+  for(int i = 0; i < (int)P.size()-1; ++i) {
     area += P[i]^(P[i+1]);
-    Point2D<int> p = P[i+1]-P[i];
+    Point2D<ll> p = P[i+1]-P[i];
     bounds += abs(__gcd(p.x, p.y));
   }
-  int inside = (abs(area) - bounds + 2)/2;
+  ll inside = (abs(area) - bounds + 2)/2;
   // Dejar el poligono como estaba antes
   P.pop_back();
   return {inside, bounds};
